pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voicebiomarkers"
version = "1.0.0"
description = "Voice Biomarker Extraction Pipeline for Cardiovascular Risk Assessment"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "CardioVoice Team"}
]
keywords = ["voice", "biomarkers", "mfcc", "praat", "cardiovascular", "machine-learning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

# =============================================================================
# DEPENDENCIES - PINNED FOR REPRODUCIBILITY
# =============================================================================
dependencies = [
    "librosa==0.10.2.post1",
    "praat-parselmouth==0.5.0",
    "noisereduce==3.0.3",
    "numpy==2.1.3",
    "scipy==1.15.2",
    "pandas==2.2.3",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyarrow>=15.0.0",  # For Parquet support
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
voicebiomarkers = "voicebiomarkers.cli:main"

[project.urls]
Repository = "https://github.com/..."

# =============================================================================
# TOOL CONFIGURATION
# =============================================================================

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]  # Line length handled by formatter

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=voicebiomarkers --cov-report=term-missing"
filterwarnings = [
    "ignore::DeprecationWarning",
]
